refactor: `EmptyData` -> `Pulse`
