+++ src/components/Icon/components.tsx
@@ -27,7 +27,6 @@ export { ReactComponent as Download } from './assets/Download.svg';
 export { ReactComponent as Earth } from './assets/Earth.svg';
 export { ReactComponent as Edit } from './assets/Edit.svg';
 export { ReactComponent as Email } from './assets/Email.svg';
-export { ReactComponent as EmptyData } from './assets/EmptyData.svg';
 export { ReactComponent as Eye } from './assets/Eye.svg';
 export { ReactComponent as EyeBlocked } from './assets/EyeBlocked.svg';
 export { ReactComponent as Faucet } from './assets/Faucet.svg';
@@ -44,6 +43,7 @@ export { ReactComponent as Page500 } from './assets/Page500.svg';
 export { ReactComponent as Plug } from './assets/Plug.svg';
 export { ReactComponent as PowerOff } from './assets/PowerOff.svg';
 export { ReactComponent as PrivateKey } from './assets/PrivateKey.svg';
+export { ReactComponent as Pulse } from './assets/Pulse.svg';
 export { ReactComponent as QRCode } from './assets/QRCode.svg';
 export { ReactComponent as Remix } from './assets/Remix.svg';
 export { ReactComponent as Settings } from './assets/Settings.svg';
