+++ src/components/Icon/urls.tsx
@@ -27,7 +27,6 @@ export { default as Download } from './assets/Download.svg';
 export { default as Earth } from './assets/Earth.svg';
 export { default as Edit } from './assets/Edit.svg';
 export { default as Email } from './assets/Email.svg';
-export { default as EmptyData } from './assets/EmptyData.svg';
 export { default as Eye } from './assets/Eye.svg';
 export { default as EyeBlocked } from './assets/EyeBlocked.svg';
 export { default as Faucet } from './assets/Faucet.svg';
@@ -44,6 +43,7 @@ export { default as Page500 } from './assets/Page500.svg';
 export { default as Plug } from './assets/Plug.svg';
 export { default as PowerOff } from './assets/PowerOff.svg';
 export { default as PrivateKey } from './assets/PrivateKey.svg';
+export { default as Pulse } from './assets/Pulse.svg';
 export { default as QRCode } from './assets/QRCode.svg';
 export { default as Remix } from './assets/Remix.svg';
 export { default as Settings } from './assets/Settings.svg';
